break off 6.0.1

+++ modules/jetty/src/test/java/org/mortbay/jetty/HttpParserTest.java
@@ -424,7 +424,7 @@ public class HttpParserTest extends TestCase
             }
         }
 
-        public void messageComplete(long contentLength)
+        public void messageComplete(int contentLength)
         {
         }
